done user service

+++ user-service/src/main/java/com/ducvu/user_service/service/AddressService.java
@@ -8,23 +8,26 @@ import com.ducvu.user_service.entity.User;
 import com.ducvu.user_service.helper.Mapper;
 import com.ducvu.user_service.repository.AddressRepository;
 import com.ducvu.user_service.repository.UserRepository;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 @Service
+@Transactional
 @RequiredArgsConstructor
 @Slf4j
 public class AddressService {
-    private AddressRepository addressRepository;
-    private UserRepository userRepository;
-    private Mapper mapper;
+    private final AddressRepository addressRepository;
+    private final UserRepository userRepository;
+    private final Mapper mapper;
 
     public AddressResponse createAddress(AddressCreateRequest request) {
         User user = userRepository.findByToken(request.getToken())
                 .orElseThrow(() -> new RuntimeException("Token invalid"));
-        Address address = mapper.toAddress(request);
 
+        Address address = mapper.toAddress(request);
+        address.setUser(user);
         addressRepository.save(address);
         user.getAddresses().add(address);
         userRepository.save(user);
@@ -32,16 +35,16 @@ public class AddressService {
         return mapper.toAddressResponse(address);
     }
 
-    public void deleteAddress(String addressId, AuthRequest request) {
+    public void deleteAddress(Integer addressId, AuthRequest request) {
         User user = userRepository.findByToken(request.getToken())
                 .orElseThrow(() -> new RuntimeException("Token invalid"));
 
-        Address address = user.getAddresses()
-                .stream()
-                .filter(a -> a.getId().equals(addressId))
-                .findFirst()
-                .orElseThrow(() -> new RuntimeException("Address not found"));
+        Address address = addressRepository.findById(addressId)
+                        .orElseThrow(() -> new RuntimeException("Address not found"));
 
+        if (!address.getUser().getId().equals(user.getId())) {
+            throw new RuntimeException("Address not found");
+        }
         addressRepository.delete(address);
         user.getAddresses().remove(address);
         userRepository.save(user);
